Migrate ThreadSafety checkers to new SAS style
