+++ src/ThreadSafety/Rules.h
@@ -1,0 +1,5 @@
+#include "ConstCastChecker.h"
+#include "ConstCastAwayChecker.h"
+#include "GlobalStaticChecker.h"
+#include "StaticLocalChecker.h"
+#include "MutableMemberChecker.h"
