+++ src/ThreadSafety/StaticLocalChecker.h
@@ -14,19 +14,25 @@
 #include <clang/StaticAnalyzer/Core/BugReporter/BugType.h>
 
 #include "SasException.h"
+#include "SasChecker.h"
 
 namespace sas
 {
-   class StaticLocalChecker : public clang::ento::Checker<clang::ento::check::ASTDecl<clang::VarDecl>>
+   namespace ThreadSafety
    {
-      mutable std::unique_ptr<clang::ento::BuiltinBug> BT;
+      class StaticLocalTraits : public CommonCheckerTraits{
+      public:
+         static constexpr const char* Name="sas.ThreadSafety.StaticLocal";
+         static constexpr const char* Description="Non-const method local static which might not be thread-safe";
+      };
+      class StaticLocalChecker : public SasChecker<StaticLocalTraits, clang::ento::check::ASTDecl<clang::VarDecl>>{
+      public:
+         void checkASTDecl(const clang::VarDecl* D, clang::ento::AnalysisManager& Mgr, clang::ento::BugReporter& BR) const;
 
-    public:
-      void checkASTDecl(const clang::VarDecl* D, clang::ento::AnalysisManager& Mgr, clang::ento::BugReporter& BR) const;
-
-    private:
-      SasException m_exception;
-   };
+      private:
+         SasException m_exception;
+      };
+   }
 }
 
 #endif
