+++ test/ThreadSafety/const_cast_away.cpp
