+++ test/ThreadSafety/global_static.cpp
