+++ test/ThreadSafety/global_static_edm.cpp
