+++ test/ThreadSafety/static_local.cpp
